Qualify the brag list a bit.
I've found several more crashes in the VM code but haven't yet reported
them; one of the difficulties in fuzz testing is finding a minimal
reproducible for a specific crash. I managed to brute force test cases
for the first two crashes by actually analysing the cores and (sort of)
understanding the problem, but that's not at all efficient.

Of course, a minimal reproducible isn't strictly necessary in order for
an expert in the code to figure out what's causing a crash. But it
speeds things up a lot: maintainers will fix bugs more quickly, allowing
me to go off and find more bugs. :)

So for the next while I'll be working on tricks to help generate small
test cases for a crash.